Extract components to own files
